+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2021-06-08, version 2.2.1
+
+- Improved handling of trailing commas.
+- Improved handling of newline delimited JSON containing commas.
+
+
 ## 2021-04-01, version 2.2.0
 
 - Implement #16: turn an escaped string containing JSON into valid JSON.